LayoutTests/platform/chromium/accessibility/aria-expanded-notifications-expected.txt

Animals
Birds
This tests that when aria-expanded is toggled, the appropriate notifications are sent out

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
Notification: AXRowCountChanged
Notification: AXRowCollapsed
Notification: AXRowCountChanged
Notification: AXRowExpanded
